Total redesign.
Now, when you hover over a link, a single svg that covers the page
is unhidden. A rect element drawn on this svg serves as the hitbox
to detect when the mouse leaves the area of the hovered :link. The
lines to other :links with the same href value are also drawn on
this one svg.

Saves adding a new svg element for every line, and calculating
bounding boxes for pairs of boxes to get its bounds right.

However, it still breaks on the sqlitetutorial.net page when not
moving the mouse slowly. Mouse pointer leaves the hitbox area
before the mouseleave listener even starts listening -> event is
missed until you re-enter and re-exit the hitbox.